Add compatibility deviations for BGP timer backwards-incomptaible change.
This change assists dependent repositories in their conversion process to support this change (https://github.com/openconfig/public/pull/604/files).

+++ feature/bgp/neighbor_test.go
@@ -263,8 +263,8 @@ func TestAugmentBGP(t *testing.T) {
 				"1.2.3.4": {
 					NeighborAddress: ygot.String("1.2.3.4"),
 					Timers: &fpoc.NetworkInstance_Protocol_Bgp_Neighbor_Timers{
-						HoldTime:          ygot.Uint16(15),
-						KeepaliveInterval: ygot.Uint16(5),
+						HoldTime:          ygot.Float64(15),
+						KeepaliveInterval: ygot.Float64(5),
 					},
 				},
 			},
@@ -278,7 +278,7 @@ func TestAugmentBGP(t *testing.T) {
 				"1.2.3.4": {
 					NeighborAddress: ygot.String("1.2.3.4"),
 					Timers: &fpoc.NetworkInstance_Protocol_Bgp_Neighbor_Timers{
-						MinimumAdvertisementInterval: ygot.Uint16(5),
+						MinimumAdvertisementInterval: ygot.Float64(5),
 					},
 				},
 			},
@@ -292,7 +292,7 @@ func TestAugmentBGP(t *testing.T) {
 				"1.2.3.4": {
 					NeighborAddress: ygot.String("1.2.3.4"),
 					Timers: &fpoc.NetworkInstance_Protocol_Bgp_Neighbor_Timers{
-						ConnectRetry: ygot.Uint16(5),
+						ConnectRetry: ygot.Float64(5),
 					},
 				},
 			},
@@ -343,7 +343,7 @@ func TestAugmentBGP_Errors(t *testing.T) {
 				"1.2.3.4": {
 					NeighborAddress: ygot.String("1.2.3.4"),
 					Timers: &fpoc.NetworkInstance_Protocol_Bgp_Neighbor_Timers{
-						MinimumAdvertisementInterval: ygot.Uint16(5),
+						MinimumAdvertisementInterval: ygot.Float64(5),
 					},
 				},
 			},
